soc: qcom: msm_power: add msm_power driver
Add driver to monitor system resources and take actions to
balance power and performance. Currently, the driver manages
system resources within the kernel but it can also be used
to notify userspace of kernel-managed system resource changes.

The driver currently monitors the frequency of all CPUs and
the GPU and limits the DDR frequency when both resources are
running below tunable thresholds.

Change-Id: Id1b446e4b3490c04b582a279496743d8c265df70
Signed-off-by: Amir Vajid <avajid@codeaurora.org>