msm: mdss: kcal: adapt to 3.18 mdss (mdp 1.7) Using the MDP 1_7 userspace structs to inject configurations of kcal to the pp_cache configs. Overriding userspace copy with memcpy when copy_from_user fails (in case of kcal interface used).
Adding a simple mdp ctl struct caching to
the kcal driver to grab ctl struct for config
calls.

Could be done nicer with direct interfacing
to userspace, leaves some space to make the kernel
crash if a third party app sends faulty configs,
but there's no security issue or degradation
of functionality with this memcpy patch.